requirements.txt
pinecone
sentence-transformers
llama-cpp-python
python-dotenv